Fixes #4450 - authorize the tasks
There are two permissions: view_foreman_tasks for viewing and
edit_foreman_tasks for controlling them.

Two default roles are pre-seeded: Tasks Reader and Tasks Manager. Also
the Anonymous role is added a permission to see user's own tasks, to
prevent the awkward moment when a user triggers a task and can't see
the progress or status.