Fixes #9508 - extract the triggering and waiting for sub plans from bulk actions
Bulk actions is just one of the use-cases for using sub-plans. This
patch extract the common code to be used for triggering external tasks
in general.

The usage is as follows: one needs to inherit it's action
from the ``Actions::ActionWithSubPlans`` and override the
``create_sub_plans`` method. The ``create_sub_plans`` will be executed
in run phase and the rest of the action will make sure that the
execution of the action will wait till the sub plan is done. It also
tracks the relation between parent and child tasks, as it already did
in bulk actions.

An example code could look like this (taken from real Katello actions):

    class SyncAsSubPlan < Actions::ActionWithSubPlans
      input_format do
        param :id, Integer
      end

      def plan(repository)
        plan_self(:id => repository.id)
      end

      def create_sub_plans
        trigger(Repository::Sync, ::Katello::Repository.find(input[:id]))
      end
    end