Refs #11895 - optimize humanized data loading
On tasks index page, call only humainzed methods that are needed.
Also, cache the humanized data on first call, as the humanization
might be relatively expensive sometimes. Also, we catch all the
exceptions when calling the humanization, as it's not critical and
the errors at humanization should not prevent the user to see other
task data.