Fixes #15624 - support for filtering data in presentation
This allows to set a `present` middleware to filter the data that are
shown to the user. To use it, one needs to define a `present`
middleware:

    class MyMiddleware < ::Dynlfow::Middleware
      def present
        action.output['password'] = '******'
        pass
      end
    end

and then load the middleware, either globally:

    initializer "my_plugin.set_dynflow_middlewares", :before => :finisher_hook do
      ForemanTasks.dynflow.config.on_init do |world|
        world.middleware.use MyMiddleware
      end
    end

or inside action (or parent action) to be used with:

   class MySecretAction < Dynflow::Action
     middleware.use MyMiddleware
   end

This should affect all the input/output data displayed to user: either
in tasks details or dynflow conosle, but should not affect the
execution.