Fixes #17914 - Improve index performance
This commit addresses a couple of inefficiencies in the controllers.
the most siginificant being a `SELECT DISTINCT foreman_tasks_tasks.*`
that was added on the resource scopes to prevent duplicates when
searching on the lock properties. The duplicates were caused by an
unneeded ext_method that created a join on the lock table, instead of
using the built in scoped search capabilities to handle has many
relations.
Another inefficiency was including id fields in non-explicit searches
making those also result in more complex queries.
In addition, the scoped search declerations were updated to use the new
:relation property instead of :in.