Fixes #18673 - Remove initialization code in favor of Dynflow
Initialization code for Rails has moved to Dynflow on 0.8.20. The only
things that should remain in Tasks with regards to that are - the web
console permissions setup and the persistence adaptor for Foreman.

The loggers, task and script to start the executor also moved to
Foreman, so the plugin should merely use them.