Fixes #25282 - Create a task for each target even if missing
We always assumed we will get all targets from the DB. If some of
the targets were missing in the DB, we didn't create the task for
it and it lead to task counters going out of sync.

With this patch we create tasks for missing targets, but pass nil
instead of the target to it.