Fixes #26560 - paused tasks notification and troubleshooting info
This path adds notifications and troubleshooting info for pauses tasks

Troubleshooting info
--------------------

A new `TroubleshootingHelpGenerator` service class has been added to
provide troubleshooting info for paused tasks. By default, it provides
generic information explaining the situation and link to troubleshooting
documentation (in case Setting[:foreman_tasks_troubleshooting_url] is
set.

The troubleshooting_url can contain macros for foreman version and
normalized task label (the label without non-alhpanumeric characters).
For example, for the following url setting:

      https://tfm.org/%{version}/tasks_troubleshooting.html#%{label}

and task 'Plugin::MyTask', the generated url would be:

      https://tfm.org/1.22/tasks_troubleshooting.html#PluginMyTask

The setting is currently not set (which means the troubleshooting link to
currently generated. We will add the setting once the page is ready in
docs.

It's also possible to extend the troubleshooting info and links on
per-action basis, by defining specific action method (see
`test/support/dummy_dynflow_action.rb` for example on how to define it.

Admin notifications
-------------------

Every time there is paused task, an admin gets a notification about
current amount of paused tasks with links to:

1. list of paused tasks
2. troubleshooting links from `TroubleshootingHelpGenerator`
  - the label for the troubleshooting url is filled with `''` in this
  case.

In case new paused task is added, the previous notification is removed
and an update notification is generated, so that the user's notification
is not overloaded with data.

Owner notification
------------------

In case the task is triggered by some user (the host-related tasks are
out of scope), the owner should the notification about the single task
being paused, with links to

1. details of the task
2. troubleshooting info for the task

How to test
-----------

1. set the `foreman_tasks_troubleshooting_url` config to some url
(as example might be
https://access.redhat.com/solutions/satellite6-tasks)
2. make some tasks to get paused
3. look at the notification drawer
4. loot the the details of the task