Fixes #26714 - Restore current user properly
In the KeepCurrentUser middleware we used to restore the current user when an
execution plan hook was being executed and then set it to nil. However with
recent addition we started running hooks even when the execution plan went into
planning.

As a timeline: Execution plan goes into planned state, hook gets triggered, the
hook cannot restore the current user, but in the end sets the current user to
nil anyway. Plan method of the entry action gets called and tries to store the
current user, however at this point the current user is already nil.