Fixes #28152 - Link tasks with user only if the user exists (#470)
Commit 2f37f56957260facecdb278fffaa1b2bd7f94eb0 changed the way how we link
tasks to users. Before we used locks, but we didn't have any constraints in
place. This could lead to a situation, where a task has a lock, which links the
task to a user, which doesn't exist anymore. If we tried to perform the
migration introduced in that commit, it would fail with foreign key violation
errors.