Fixes #29203 - Fix an edge case with excessive progress precision
Before this patch we took a float in the range from 0 to 1, turned it into a
string with desired precision and then multiplied it with 100 to get the
percentage. However, this didn't really work for all inputs, because of the well
known issues when multiplying floats.

> x = 0.004545454545454546
0.004545454545454546

> x.toFixed(4) * 100
0.44999999999999996

> parseFloat((x * 100).toFixed(2))
0.45

(cherry picked from commit 0eb1e2ed1ce29af20d9f9030ad68f47cbd5d020d)