Fixes #30161 - Log a deprecation warning if per-step current user is used
Starting with the previous commit current user is set for the entire finalize
phase and is based on the user saved for the entry action of the task. This
deprecates the old behavior of setting and unsetting current user per step. If
we detect that an action relies on the old behavior by overriding the user set
for the entire phase, we emit a deprecation warning.