Fixes #32047 - Match proxy actions using task_id (#617)
When we're retrieving outputs of proxy actions we get back a serialized
execution plan, walk its actions and look for an action with either expected
action_class or operation name. This is somewhat flaky.

With this change we only check one field in the actions' input and that is
task_id. First action found with matching value is considered the source of
output for the current task.