Fixes #32729 - reload setting prior run
In Foreman we've switched to in-memory setting cache, that needs to be
reloaded on demand, Rails is doing that on every request, for dynflow
actions we need to do it on every phase.

It does that only if the Foreman version is 2.5 or higher, as for older
versions this would fail.