Fixes #32132 - Make tasks counts match when using dashboard
We had two ways of filtering tasks in the controllers. One relied on the search
query passed as search parameter, the other relied on a couple of special
parameters. If any of the special parameters were provided, we performed the
search *before* we passed the scope to the helper which does the actual search
based searching. Because of this total and subtotal in the response did not
match what was shown.

This change throws away the special parameters handling from the backend and
makes it frontend's responsibility to assemble proper search query.

(cherry picked from commit 9119db11280949fc464b449301a960d667602f65)