Fixes #35204 - Add an option to pre-fetch proxy task statuses
The usage of continuous output, proxy action and delegated action makes
it somewhat cumbersome to load the data in bulk, but still have it
processed the way we're used to. It would require a much bigger
refactoring.

The current solution allows preloading task statuses into rails cache,
from where they can be the read by the regular continuous output
machinery.