Refs #35328 - Workaround for serialization issues
Apparently the exception contains both the response and the original
request through the response. If I recall correctly, the body of the
response is read on demand and the actual read is synchronized so it is
thread safe. The monitor used for this synchronization cannot be
serialized, pass the boundaries of a process and then be instantiated
again. It is ugly, but so far we only care about the exception's class
and its backtrace.