Fixes #35328 - Propagate errors when remote parent task fails (#691)
* Fixes #35328 - Propagate errors when remote parent task fails

Before we notified the local child actions that their remote counterpart
has failed, but we didn't include any details. This made the local
actions ask the smart proxy for status, get back a 404, which the local
actions did not expect.

If we know that the remote parent failed during triggering, it is clear
there will not be any remote tasks to ask about, so we can fail straight
away, without having to ask for status.

* Refs #35328 - Properly deal with tasks disappearing from the proxy

It can still happen that we can try asking for details about a remote
task which no longer exists.

* Refs #35328 - Workaround for serialization issues

Apparently the exception contains both the response and the original
request through the response. If I recall correctly, the body of the
response is read on demand and the actual read is synchronized so it is
thread safe. The monitor used for this synchronization cannot be
serialized, pass the boundaries of a process and then be instantiated
again. It is ugly, but so far we only care about the exception's class
and its backtrace.