Fixes #35860 - Fix untranslatable strings
It is never valid to use __(``) since string extraction relies on static
strings. String interpolation breaks this, so it's best to avoid it even
if the string contained is static.

This also fixes a string where interpolation should happen. This allows
translations to move around the word order. Ideally they would actually
properly use pluralization so that it would be "select 1 task" instead
of "select 1 tasks". That would also allow other languages to properly
use plurals, but that's left as a future improvement.