Refs #36848 - Disable mail setting hook when in seeds
Returning early if we're seeding should be safe, under the assumtion
there is no admin user that gets seeded after the notification itself,
because the notification attaches itself to admins explicitly when it is
created and does not rely on the before verification callback.

It should now survive when the org admin role does not exist.

(cherry picked from commit 165cb032433f21af3d2715a450a0e08af93e56af)